# static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Event Management API",
    "description": "This is a backend for event management application",
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081/api"
    }
  ],
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "login using credentials",
        "operationId": "login",
        "requestBody": {
          "description": "credentials",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BadRequestResponse"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid Credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error most likely due to service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "sign up as new user",
        "operationId": "register",
        "requestBody": {
          "description": "user's basic information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BadRequestResponse"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal error most likely due to service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/check": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "check",
        "summary": "check if access token is still valid",
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ServerErrorResponse"
                    },
                    {
                      "properties": {
                        "messages": {
                          "example": "repository connection lost"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "9dc2391b-595e-4a9a-aefb-75dbe18fe5ca"
          },
          "username": {
            "type": "string",
            "example": "john_doe"
          },
          "first_name": {
            "type": "string",
            "example": "John"
          },
          "last_name": {
            "type": "string",
            "example": "Doe"
          },
          "role": {
            "type": "string",
            "example": "organizer"
          }
        }
      },
      "RestAPIResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "ErrorAPIResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/RestAPIResponse"
          },
          {
            "properties": {
              "success": {
                "example": false
              },
              "code": {
                "type": "string",
                "example": "AUTH_INVALID_CREDENTIALS"
              },
              "messages": {
                "oneOf": [
                  {
                    "type": "string",
                    "example": "more information about the error..."
                  },
                  {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "more information about the error..."
                    }
                  }
                ]
              }
            }
          }
        ]
      },
      "BadRequestResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ErrorAPIResponse"
          },
          {
            "properties": {
              "code": {
                "type": "string",
                "example": "BAD_REQUEST"
              },
              "messages": {
                "example": [
                  "validation error"
                ]
              }
            }
          }
        ]
      },
      "ServerErrorResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ErrorAPIResponse"
          },
          {
            "properties": {
              "code": {
                "example": "INTERNAL_SERVER_ERROR"
              },
              "messages": {
                "example": [
                  "Service is unavailable"
                ]
              }
            }
          }
        ]
      },
      "UnauthorizedResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ErrorAPIResponse"
          },
          {
            "properties": {
              "code": {
                "example": "AUTH_INVALID_CREDENTIALS"
              },
              "messages": {
                "example": [
                  "invalid credentials"
                ]
              }
            }
          }
        ]
      },
      "NotFoundErrorResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ErrorAPIResponse"
          },
          {
            "properties": {
              "code": {
                "example": "NOT_FOUND"
              },
              "messages": {
                "example": [
                  "user with id 9dc2391b-595e-4a9a-aefb-75dbe18fe5ca does not exist"
                ]
              }
            }
          }
        ]
      },
      "CheckResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/RestAPIResponse"
          },
          {
            "properties": {
              "data": {
                "$ref": "#/components/schemas/UserInfo"
              }
            }
          }
        ]
      },
...
      "UserData": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "curiousdev@bedevs.org"
          },
          "first_name": {
            "type": "string",
            "example": "john"
          },
          "last_name": {
            "type": "string",
            "example": "doe"
          },
          "username": {
            "type": "string",
            "example": "john_doe"
          }
        }
      },
      "RegisterRequest": {
        "required": [
          "email",
          "username",
          "password"
        ],
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/UserData"
          },
          {
            "properties": {
              "password": {
                "type": "string",
                "example": "CuriousDevs2024!",
                "description": "password must be at least 6 characters long and alphanumeric"
              }
            }
          }
        ]
      },
      "RegisterResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/RestAPIResponse"
          },
          {
            "properties": {
              "data": {
                "type": "string",
                "example": "successfully signed up"
              }
            }
          }
        ]
      },
      "LoginResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/RestAPIResponse"
          },
          {
            "properties": {
              "data": {
                "$ref": "#/components/schemas/LoginSuccessDTO"
              }
            }
          }
        ]
      },
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "curiousdev@bedevs.org"
          },
          "password": {
            "type": "string",
            "example": "CuriousDevs2024!"
          }
        }
      },
      "LoginSuccessDTO": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserInfo"
          },
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
          }
        }
      },
...
    },
    "securitySchemes": {
      "tokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}